Continued config refactor
